re-allocate codes for calculate psnr
